flaskblog/requirements.txt
Flask==0.10.1
PyMySQL==0.6.7
Flask-SQLAlchemy==2.1
Flask-Script==2.0.5
Flask-WTF==0.12
Flask-Migrate==1.6.0
py-bcrypt==0.4
python-slugify==1.2.1
Flask-Markdown==0.3
Flask-Uploads==0.2.0
celery=4.0.0
redis=3.2.0
